// packages/test/generatedCode/policies/roleManagementPolicies/item/unifiedRoleManagementPolicyItemRequestBuilderGetRequestConfiguration.ts
import {UnifiedRoleManagementPolicyItemRequestBuilderGetQueryParameters} from './unifiedRoleManagementPolicyItemRequestBuilderGetQueryParameters';
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface UnifiedRoleManagementPolicyItemRequestBuilderGetRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string>;
    /** Request options */
    options?: RequestOption[];
    /** Request query parameters */
    queryParameters?: UnifiedRoleManagementPolicyItemRequestBuilderGetQueryParameters;
}